Created helper to save D3JS graphs as PNG. Added button to save graphs as PNG to each chart.
